Remove duplicate type.
