+++ enterprise/cmd/frontend/internal/executorqueue/queues/codeintel/queue.go
@@ -6,6 +6,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executorqueue/handler"
 	apiclient "github.com/sourcegraph/sourcegraph/enterprise/internal/executor"
 	store "github.com/sourcegraph/sourcegraph/internal/codeintel/stores/dbstore"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/types"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
@@ -14,7 +15,7 @@ import (
 
 func QueueOptions(db database.DB, accessToken func() string, observationContext *observation.Context) handler.QueueOptions {
 	recordTransformer := func(ctx context.Context, record workerutil.Record) (apiclient.Job, error) {
-		return transformRecord(record.(store.Index), accessToken())
+		return transformRecord(record.(types.Index), accessToken())
 	}
 
 	return handler.QueueOptions{
