+++ enterprise/cmd/frontend/internal/executorqueue/queues/codeintel/transform.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 
 	apiclient "github.com/sourcegraph/sourcegraph/enterprise/internal/executor"
-	store "github.com/sourcegraph/sourcegraph/internal/codeintel/stores/dbstore"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/types"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
 )
 
@@ -14,7 +14,7 @@ const defaultOutfile = "dump.lsif"
 const uploadRoute = "/.executors/lsif/upload"
 const schemeExecutorToken = "token-executor"
 
-func transformRecord(index store.Index, accessToken string) (apiclient.Job, error) {
+func transformRecord(index types.Index, accessToken string) (apiclient.Job, error) {
 	dockerSteps := make([]apiclient.DockerStep, 0, len(index.DockerSteps)+2)
 	for _, dockerStep := range index.DockerSteps {
 		dockerSteps = append(dockerSteps, apiclient.DockerStep{
