+++ internal/codeintel/sharedresolvers/types.go
@@ -1,0 +1,9 @@
+package sharedresolvers
+
+import "github.com/sourcegraph/sourcegraph/internal/codeintel/types"
+
+type IndexesWithRepositoryNamespace struct {
+	Root    string
+	Indexer string
+	Indexes []types.Index
+}
